separate inspect-buildpack display structs for BOM vs non-BOM.
- clean up
